fix: fix all db locked errors (#232)
Description
---
- uses a single shared Sqlite connection per database
- updates to the latest nightly (GAT stabilized)
- run database migrations once on startup
- remove circleci manifest

Motivation and Context
---
No more "database is locked" errors

How Has This Been Tested?
---
Manually

+++ dan_layer/core/src/services/leader_strategy.rs
@@ -84,7 +84,7 @@ impl<TAddr: NodeAddressable> LeaderStrategy<TAddr> for PayloadSpecificLeaderStra
             .finalize()
             .into();
         let hash = u32::from_le_bytes([hash[0], hash[1], hash[2], hash[3]]);
-        let first = (hash % committee.members.len() as u32) as u32;
+        let first = hash % committee.members.len() as u32;
         (first + round) % committee.members.len() as u32
     }
 }
